requirements.txt
PyYAML~=6.0.2
pandas~=2.2.3
numpy~=2.1.3
prophet~=1.1.6
scikit-learn~=1.5.2
matplotlib~=3.9.2
joblib~=1.4.2
seaborn~=0.13.2
optuna~=4.0.0